// manual-scripts/shared/src/types.ts
export enum Marketplace {
    MagicEdenV1 = 'MagicEdenV1',
    MagicEdenV2 = 'MagicEdenV2',
};

export class MarketplaceTransactionEntityData {
    buyerWalletAccount: StringPublicKey;
    buyerPdaAccount: StringPublicKey;
    sellerWalletAccount: StringPublicKey;
    sellerPdaAccount: StringPublicKey;
    tokenMintAccount: StringPublicKey;
    created: number;
    signature: TransactionSignature;
    marketplace: Marketplace;
    pdaData: string;
    price: number;

    constructor(
        buyerWalletAccount: StringPublicKey,
        buyerPdaAccount: StringPublicKey,
        sellerWalletAccount: StringPublicKey,
        sellerPdaAccount: StringPublicKey,
        tokenMintAccount: StringPublicKey,
        created: number,
        signature: TransactionSignature,
        marketplace: Marketplace,
        pdaData: string,
        price: number,
    )
    {
        this.buyerWalletAccount = buyerWalletAccount;
        this.buyerPdaAccount = buyerPdaAccount;
        this.sellerWalletAccount = sellerWalletAccount;
        this.sellerPdaAccount = sellerPdaAccount;
        this.tokenMintAccount = tokenMintAccount;
        this.created = created;
        this.signature = signature;
        this.marketplace = marketplace;
        this.pdaData = pdaData;
        this.price = price;
    }
};

export class MarketplaceTransactionEntityDataWithType extends MarketplaceTransactionEntityData {
    type: EntityType;
};

export interface CollectionMetaData {
    name?: string;
    symbol?: string;
    elementCount?: number;
    updateAuthority?: string;
    creator?: string;
    description?: string;
    family?: string;
    image?: string;
}

export interface CollectionAttributeData {
    attributes: CollectionAttribute[];
}

export interface CollectionPriceData {
    volume: CollectionVolume;
    floorPrice: CollectionPrice;
    ceilingPrice: CollectionPrice;
}

export interface CollectionAttribute {
    name?: string;
    key?: string;
    values: CollectionAttributeValue[];
}
// ...
export enum EntityType {
    BidTransaction = 'bidTransaction',
    AskTransaction = 'askTransaction',
    CancelBidTransaction = 'cancelBidTransaction',
    CancelAskTransaction = 'cancelAskTransaction',
    ExecuteSaleTransaction = 'executeSaleTransaction',
    MarketplaceTransactionForNFT = 'marketplaceTransaction4NFT',
    MarketplaceTransactionForWallet = 'marketplaceTransaction4Wallet',

    BidOrderState = 'bidOrderState',
    AskOrderState = 'askOrderState',

    UserRefreshTokenEntity = 'userRefreshToken',
    NonceEntity = 'nonce',

    MetaplexOnChainMetadata = 'MetaplexOnChainMetadata',
    MintAddress = 'mintAddress',
    NFTOffchainMetadata = 'NFToffchainMetadata',
    MetaplexOffchainMetadata = 'MetaplexOffchainMetadata',
    Trigger = 'trigger',
    CensorTag = "censorTag",
    CollectionID = "collectionID",
    UserFlag = "userFlag",
    StaleFlag = "staleFlag",

    CollectionAttributeData = 'collectionAttributeData',
    CollectionMetaData = 'collectionMetaData',
    CollectionPriceData = 'collectionPriceData',

    NFT4Collection = 'NFT4Collection',
    Collection4NFT = 'collection4NFT',
    PlaydustCollection = 'playdustCollection',
    Transaction = 'transaction',

    MarketData4Collection = 'marketData4Collection',
    OffchainMetadata4Collection = 'offchainMetadata4Collection',
    NFTRarity4Collection = 'NFTRarity4Collection',

    NFTRarity = 'NFTRarity',
}
// ...
export type StringPublicKey = string;
// ...
export interface Entity {
    globalId: string;
    id?: string;
    type?: EntityType;
    primaryEntity?: string;
    createdAt?: Date;
    updatedAt: Date;
    // FIXME: Use a generic type here
    data?: any;
};
// ...
export interface Transaction {
    /** The transaction message */
    message: Message;
    /** The transaction signatures */
    signatures: string[];
}

interface ByMarketplaceCollectionData<T> {
    byMarketplace?: Map<Marketplace, T>;
    global: T;
};

export interface VolumeData {
    d1: number;
    d7: number;
    d30: number;
    total: number;
};

export type CollectionPrice = ByMarketplaceCollectionData<number>;
export type CollectionVolume = ByMarketplaceCollectionData<VolumeData>;